add constant file config
