+++ src/api/subscriptions/subscriptions.controller.ts
@@ -1,9 +1,11 @@
 import { Body, Controller, Post, UseInterceptors } from '@nestjs/common';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import { SubscriptionsService } from './subscriptions.service';
-import { CreateSubscriptionDto } from 'src/models/subscription/subscription.dto';
+import {
+  CreateSubscriptionDto,
+  MergedSubscriptions,
+} from 'src/models/subscription/subscription.dto';
 import { LoggingInterceptor } from 'src/common/interceptors/logger.interceptor';
-import { MergedSubscriptions } from './subscriptions.service';
 
 @ApiTags('Subscriptions')
 @UseInterceptors(LoggingInterceptor)
