+++ src/models/activity/activity.providers.ts
@@ -1,10 +1,11 @@
 import { DataSource } from 'typeorm';
+import { ACTIVITY_REPOSITORY, DWC_MYSQL_CONNECTION } from '../constants';
 import { Activity } from './activity.dwc.entity';
 
 export const activityProviders = [
   {
-    provide: 'ACTIVITY_REPOSITORY',
+    provide: ACTIVITY_REPOSITORY,
     useFactory: (dataSource: DataSource) => dataSource.getRepository(Activity),
-    inject: ['DWC_MYSQL_CONNECTION'],
+    inject: [DWC_MYSQL_CONNECTION],
   },
 ];
