+++ src/models/constants.ts
@@ -1,0 +1,8 @@
+//#REPOSITORY
+export const SUBSCRIPTION_REPOSITORY = 'SUBSCRIPTION_REPOSITORY';
+export const CUSTOMER_REPOSITORY = 'CUSTOMER_REPOSITORY';
+export const ACTIVITY_REPOSITORY = 'ACTIVITY_REPOSITORY';
+
+//#CONNECTION
+export const DWC_MYSQL_CONNECTION = 'DWC_MYSQL_CONNECTION';
+export const PROXI_MYSQL_CONNECTION = 'PROXI_MYSQL_CONNECTION';
