+++ src/models/customer/customer.providers.ts
@@ -1,10 +1,11 @@
 import { DataSource } from 'typeorm';
+import { CUSTOMER_REPOSITORY, PROXI_MYSQL_CONNECTION } from '../constants';
 import { Customer } from './customer.proxi.entity';
 
 export const customerProviders = [
   {
-    provide: 'CUSTOMER_REPOSITORY',
+    provide: CUSTOMER_REPOSITORY,
     useFactory: (dataSource: DataSource) => dataSource.getRepository(Customer),
-    inject: ['PROXI_MYSQL_CONNECTION'],
+    inject: [PROXI_MYSQL_CONNECTION],
   },
 ];
